requirements.txt
aiohttp==3.11.11
numpy<2
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==11.1.0
python-dotenv==1.0.1
Requests==2.32.3
rich==13.9.4
tqdm==4.65.0
beautifulsoup4==4.12.3
PyYAML==6.0.2
pydantic==2.11.3
py-cord==2.6.1
discord.py==2.5.2
